+++ README.md
@@ -1,3 +1,97 @@
 # sudokuScala
 
-A scala sudoku solver.  The program first tries to solve the puzzle with sudoku logic.  If that fails, it uses a backtracking algorithm to find a solution (if one exists).
+A scala sudoku solver.  
+
+## Usage
+
+```
+alexpavlakis$ scala sudokuSolver.scala
+
+easy sudoku
++ - - - + - - - + - - - +
+| 2 4   |   8   | 5     | 
+| 1   6 |     5 |   2   | 
+|   8 5 | 2 7   |       | 
++ - - - + - - - + - - - +
+|       | 5   2 |   7   | 
+|     2 |       | 1     | 
+|   1   | 3   4 |       | 
++ - - - + - - - + - - - +
+|       |   5 7 | 4 6   | 
+|   2   | 9     | 7   8 | 
+|     4 |   2   |   1 9 | 
++ - - - + - - - + - - - +
+Elapsed time: 0.0264 seconds
++ - - - + - - - + - - - +
+| 2 4 7 | 1 8 9 | 5 3 6 | 
+| 1 9 6 | 4 3 5 | 8 2 7 | 
+| 3 8 5 | 2 7 3 | 9 4 1 | 
++ - - - + - - - + - - - +
+| 8 6 3 | 5 1 2 | 9 7 4 | 
+| 4 5 2 | 7 9 8 | 1 9 3 | 
+| 7 1 9 | 3 6 4 | 2 8 5 | 
++ - - - + - - - + - - - +
+| 9 3 9 | 8 5 7 | 4 6 2 | 
+| 6 2 1 | 9 4 3 | 7 5 8 | 
+| 5 7 4 | 6 2 6 | 3 1 9 | 
++ - - - + - - - + - - - +
+
+hard sudoku
++ - - - + - - - + - - - +
+| 8     |       | 2     | 
+| 7     |   8   |   5 9 | 
+|       | 9   2 |     6 | 
++ - - - + - - - + - - - +
+| 5     |   7   |   1   | 
+|     4 |     8 |       | 
+|   8   |     6 |       | 
++ - - - + - - - + - - - +
+|       |       |   2   | 
+|       |   1   |   3   | 
+| 3 7   |   6   |       | 
++ - - - + - - - + - - - +
+Elapsed time: 0.0034 seconds
++ - - - + - - - + - - - +
+| 8 6 9 | 7 3 5 | 2 4 1 | 
+| 7 4 2 | 6 8 1 | 3 5 9 | 
+| 1 5 3 | 9 4 2 | 8 7 6 | 
++ - - - + - - - + - - - +
+| 5 3 6 | 2 7 9 | 4 1 8 | 
+| 9 1 4 | 3 5 8 | 7 6 2 | 
+| 2 8 7 | 1 5 6 | 5 9 3 | 
++ - - - + - - - + - - - +
+| 4 6 5 | 8 9 3 | 1 2 7 | 
+| 4 2 8 | 8 1 7 | 6 3 5 | 
+| 3 7 1 | 5 6 4 | 9 8 5 | 
++ - - - + - - - + - - - +
+
+evil sudoku
++ - - - + - - - + - - - +
+|       |       |   1 2 | 
+|       |       |     3 | 
+|     2 | 3     | 4     | 
++ - - - + - - - + - - - +
+|     1 | 8     |     5 | 
+|   6   |   7   | 8     | 
+|       |     9 |       | 
++ - - - + - - - + - - - +
+|     8 | 5     |       | 
+| 9     |   4   | 5     | 
+| 4 7   |     6 |       | 
++ - - - + - - - + - - - +
+Elapsed time: 0.0046 seconds
++ - - - + - - - + - - - +
+| 3 4 9 | 7 9 8 | 6 1 2 | 
+| 1 8 6 | 4 1 2 | 7 5 3 | 
+| 7 5 2 | 3 6 5 | 4 9 8 | 
++ - - - + - - - + - - - +
+| 2 9 1 | 8 3 4 | 7 6 5 | 
+| 5 6 4 | 2 7 1 | 8 3 9 | 
+| 8 3 7 | 6 5 9 | 1 2 4 | 
++ - - - + - - - + - - - +
+| 6 1 8 | 5 2 3 | 9 4 7 | 
+| 9 2 3 | 1 4 7 | 5 8 6 | 
+| 4 7 5 | 9 8 6 | 3 2 1 | 
++ - - - + - - - + - - - +
+
+```
