+++ sudokuPuzzles.scala
@@ -1,85 +1,3 @@
-// Board metadata
-var rows = List(1, 2, 3, 4, 5, 6, 7, 8, 9, 
-				1, 2, 3, 4, 5, 6, 7, 8, 9, 
-				1, 2, 3, 4, 5, 6, 7, 8, 9,
-				1, 2, 3, 4, 5, 6, 7, 8, 9,
-				1, 2, 3, 4, 5, 6, 7, 8, 9,
-				1, 2, 3, 4, 5, 6, 7, 8, 9,
-				1, 2, 3, 4, 5, 6, 7, 8, 9, 
-				1, 2, 3, 4, 5, 6, 7, 8, 9, 
-				1, 2, 3, 4, 5, 6, 7, 8, 9)
-var cols = rows.sorted
-var boxs = List(1, 1, 1, 4, 4, 4, 7, 7, 7, 
-				1, 1, 1, 4, 4, 4, 7, 7, 7, 
-				1, 1, 1, 4, 4, 4, 7, 7, 7, 
-				2, 2, 2, 5, 5, 5, 8, 8, 8, 
-				2, 2, 2, 5, 5, 5, 8, 8, 8, 
-				2, 2, 2, 5, 5, 5, 8, 8, 8, 
-				3, 3, 3, 6, 6, 6, 9, 9, 9, 
-				3, 3, 3, 6, 6, 6, 9, 9, 9, 
-				3, 3, 3, 6, 6, 6, 9, 9, 9)
-
-
-
-// Easy Sudoku
-var vals = List(2, 4, 0, 0, 8, 0, 5, 0, 0, 
-				1, 0, 6, 0, 0, 5, 0, 2, 0, 
-				0, 8, 5, 2, 7, 0, 0, 0, 0, 
-				0, 0, 0, 5, 0, 2, 0, 7, 0, 
-				0, 0, 2, 0, 0, 0, 1, 0, 0, 
-				0, 1, 0, 3, 0, 4, 0, 0, 0, 
-				0, 0, 0, 0, 5, 7, 4, 6, 0, 
-				0, 2, 0, 9, 0, 0, 7, 0, 8, 
-				0, 0, 4, 0, 2, 0, 0, 1, 9)
-
-var easySudoku = Array.ofDim[Int](81, 4)
-for(i <- 0.to(80)) {
-  easySudoku(i)(0) = vals(i)
-  easySudoku(i)(1) = rows(i)
-  easySudoku(i)(2) = cols(i)
-  easySudoku(i)(3) = boxs(i)
-}
-
-
-
-
-// Hard Sudoku
-var vals = List(8, 0, 0, 0, 0, 0, 2, 0, 0, 
-				7, 0, 0, 0, 8, 0, 0, 5, 9, 
-				0, 0, 0, 9, 0, 2, 0, 0, 6, 
-				5, 0, 0, 0, 7, 0, 0, 1, 0, 
-				0, 0, 4, 0, 0, 8, 0, 0, 0, 
-				0, 8, 0, 0, 0, 6, 0, 0, 0, 
-				0, 0, 0, 0, 0, 0, 0, 2, 0, 
-				0, 0, 0, 0, 1, 0, 0, 3, 0, 
-				3, 7, 0, 0, 6, 0, 0, 0, 0)
-
-var hardSudoku = Array.ofDim[Int](81, 4)
-for(i <- 0.to(80)) {
-  hardSudoku(i)(0) = vals(i)
-  hardSudoku(i)(1) = rows(i)
-  hardSudoku(i)(2) = cols(i)
-  hardSudoku(i)(3) = boxs(i)
-}
-
-
-
-
-// Evil Sudoku
-var vals = List(0, 0, 0, 0, 0, 0, 0, 1, 2, 
-				0, 0, 0, 0, 0, 0, 0, 0, 3, 
-				0, 0, 2, 3, 0, 0, 4, 0, 0, 
-				0, 0, 1, 8, 0, 0, 0, 0, 5, 
-				0, 6, 0, 0, 7, 0, 8, 0, 0, 
-				0, 0, 0, 0, 0, 9, 0, 0, 0, 
-				0, 0, 8, 5, 0, 0, 0, 0, 0, 
-				9, 0, 0, 0, 4, 0, 5, 0, 0, 
-				4, 7, 0, 0, 0, 6, 0, 0, 0)
-
-var evilSudoku = Array.ofDim[Int](81, 4)
-for(i <- 0.to(80)) {
-  evilSudoku(i)(0) = vals(i)
-  evilSudoku(i)(1) = rows(i)
-  evilSudoku(i)(2) = cols(i)
-  evilSudoku(i)(3) = boxs(i)
-}
+var easy_sudoku = ListBuffer(2,4,0,0,8,0,5,0,0,1,0,6,0,0,5,0,2,0,0,8,5,2,7,0,0,0,0,0,0,0,5,0,2,0,7,0,0,0,2,0,0,0,1,0,0,0,1,0,3,0,4,0,0,0,0,0,0,0,5,7,4,6,0,0,2,0,9,0,0,7,0,8,0,0,4,0,2,0,0,1,9)
+var hard_sudoku = ListBuffer(8,0,0,0,0,0,2,0,0,7,0,0,0,8,0,0,5,9,0,0,0,9,0,2,0,0,6,5,0,0,0,7,0,0,1,0,0,0,4,0,0,8,0,0,0,0,8,0,0,0,6,0,0,0,0,0,0,0,0,0,0,2,0,0,0,0,0,1,0,0,3,0,3,7,0,0,6,0,0,0,0)
+var evil_sudoku = ListBuffer(0,0,0,0,0,0,0,1,2,0,0,0,0,0,0,0,0,3,0,0,2,3,0,0,4,0,0,0,0,1,8,0,0,0,0,5,0,6,0,0,7,0,8,0,0,0,0,0,0,0,9,0,0,0,0,0,8,5,0,0,0,0,0,9,0,0,0,4,0,5,0,0,4,7,0,0,0,6,0,0,0)
