+++ sudokuSolver.scala
@@ -1,223 +1,221 @@
-// This program solves sudoku puzzles structured as arrays of arrays, 
-// with 0 for unknown elements.  The program first tries to solve the
-// puzzle with sudoku logic.  If that fails, it employs a backtracking
-// algorithm to find a solution (if one exists)
+// This program solves sudoku puzzles structured as ListBuffers, 
+// with 0 for unknown elements with a fast backtracking algorithm
+
 import scala.collection.mutable.ListBuffer
 
 object Sudoku {
 
-	// Possibilities for a sudoku
-	val nums = List(1, 2, 3, 4, 5, 6, 7, 8, 9)
+	// The indices of peers for each element (81x20)
+	val peers = List( List(1,2,3,4,5,6,7,8,9,10,11,18,19,20,27,36,45,54,63,72),
+			          List(0,2,3,4,5,6,7,8,9,10,11,18,19,20,28,37,46,55,64,73),
+			          List(0,1,3,4,5,6,7,8,9,10,11,18,19,20,29,38,47,56,65,74),
+			          List(0,1,2,4,5,6,7,8,12,13,14,21,22,23,30,39,48,57,66,75),
+			          List(0,1,2,3,5,6,7,8,12,13,14,21,22,23,31,40,49,58,67,76),
+			          List(0,1,2,3,4,6,7,8,12,13,14,21,22,23,32,41,50,59,68,77),
+			          List(0,1,2,3,4,5,7,8,15,16,17,24,25,26,33,42,51,60,69,78),
+			          List(0,1,2,3,4,5,6,8,15,16,17,24,25,26,34,43,52,61,70,79),
+			          List(0,1,2,3,4,5,6,7,15,16,17,24,25,26,35,44,53,62,71,80),
+			          List(0,1,2,10,11,12,13,14,15,16,17,18,19,20,27,36,45,54,63,72),
+			          List(0,1,2,9,11,12,13,14,15,16,17,18,19,20,28,37,46,55,64,73),
+			          List(0,1,2,9,10,12,13,14,15,16,17,18,19,20,29,38,47,56,65,74),
+			          List(3,4,5,9,10,11,13,14,15,16,17,21,22,23,30,39,48,57,66,75),
+			          List(3,4,5,9,10,11,12,14,15,16,17,21,22,23,31,40,49,58,67,76),
+			          List(3,4,5,9,10,11,12,13,15,16,17,21,22,23,32,41,50,59,68,77),
+			          List(6,7,8,9,10,11,12,13,14,16,17,24,25,26,33,42,51,60,69,78),
+			          List(6,7,8,9,10,11,12,13,14,15,17,24,25,26,34,43,52,61,70,79),
+			          List(6,7,8,9,10,11,12,13,14,15,16,24,25,26,35,44,53,62,71,80),
+			          List(0,1,2,9,10,11,19,20,21,22,23,24,25,26,27,36,45,54,63,72),
+			          List(0,1,2,9,10,11,18,20,21,22,23,24,25,26,28,37,46,55,64,73),
+			          List(0,1,2,9,10,11,18,19,21,22,23,24,25,26,29,38,47,56,65,74),
+			          List(3,4,5,12,13,14,18,19,20,22,23,24,25,26,30,39,48,57,66,75),
+			          List(3,4,5,12,13,14,18,19,20,21,23,24,25,26,31,40,49,58,67,76),
+			          List(3,4,5,12,13,14,18,19,20,21,22,24,25,26,32,41,50,59,68,77),
+			          List(6,7,8,15,16,17,18,19,20,21,22,23,25,26,33,42,51,60,69,78),
+			          List(6,7,8,15,16,17,18,19,20,21,22,23,24,26,34,43,52,61,70,79),
+			          List(6,7,8,15,16,17,18,19,20,21,22,23,24,25,35,44,53,62,71,80),
+			          List(0,9,18,28,29,30,31,32,33,34,35,36,37,38,45,46,47,54,63,72),
+			          List(1,10,19,27,29,30,31,32,33,34,35,36,37,38,45,46,47,55,64,73),
+			          List(2,11,20,27,28,30,31,32,33,34,35,36,37,38,45,46,47,56,65,74),
+			          List(3,12,21,27,28,29,31,32,33,34,35,39,40,41,48,49,50,57,66,75),
+			          List(4,13,22,27,28,29,30,32,33,34,35,39,40,41,48,49,50,58,67,76),
+			          List(5,14,23,27,28,29,30,31,33,34,35,39,40,41,48,49,50,59,68,77),
+			          List(6,15,24,27,28,29,30,31,32,34,35,42,43,44,51,52,53,60,69,78),
+			          List(7,16,25,27,28,29,30,31,32,33,35,42,43,44,51,52,53,61,70,79),
+			          List(8,17,26,27,28,29,30,31,32,33,34,42,43,44,51,52,53,62,71,80),
+			          List(0,9,18,27,28,29,37,38,39,40,41,42,43,44,45,46,47,54,63,72),
+			          List(1,10,19,27,28,29,36,38,39,40,41,42,43,44,45,46,47,55,64,73),
+			          List(2,11,20,27,28,29,36,37,39,40,41,42,43,44,45,46,47,56,65,74),
+			          List(3,12,21,30,31,32,36,37,38,40,41,42,43,44,48,49,50,57,66,75),
+			          List(4,13,22,30,31,32,36,37,38,39,41,42,43,44,48,49,50,58,67,76),
+			          List(5,14,23,30,31,32,36,37,38,39,40,42,43,44,48,49,50,59,68,77),
+			          List(6,15,24,33,34,35,36,37,38,39,40,41,43,44,51,52,53,60,69,78),
+			          List(7,16,25,33,34,35,36,37,38,39,40,41,42,44,51,52,53,61,70,79),
+			          List(8,17,26,33,34,35,36,37,38,39,40,41,42,43,51,52,53,62,71,80),
+			          List(0,9,18,27,28,29,36,37,38,46,47,48,49,50,51,52,53,54,63,72),
+			          List(1,10,19,27,28,29,36,37,38,45,47,48,49,50,51,52,53,55,64,73),
+			          List(2,11,20,27,28,29,36,37,38,45,46,48,49,50,51,52,53,56,65,74),
+			          List(3,12,21,30,31,32,39,40,41,45,46,47,49,50,51,52,53,57,66,75),
+			          List(4,13,22,30,31,32,39,40,41,45,46,47,48,50,51,52,53,58,67,76),
+			          List(5,14,23,30,31,32,39,40,41,45,46,47,48,49,51,52,53,59,68,77),
+			          List(6,15,24,33,34,35,42,43,44,45,46,47,48,49,50,52,53,60,69,78),
+			          List(7,16,25,33,34,35,42,43,44,45,46,47,48,49,50,51,53,61,70,79),
+			          List(8,17,26,33,34,35,42,43,44,45,46,47,48,49,50,51,52,62,71,80),
+			          List(0,9,18,27,36,45,55,56,57,58,59,60,61,62,63,64,65,72,73,74),
+			          List(1,10,19,28,37,46,54,56,57,58,59,60,61,62,63,64,65,72,73,74),
+			          List(2,11,20,29,38,47,54,55,57,58,59,60,61,62,63,64,65,72,73,74),
+			          List(3,12,21,30,39,48,54,55,56,58,59,60,61,62,66,67,68,75,76,77),
+			          List(4,13,22,31,40,49,54,55,56,57,59,60,61,62,66,67,68,75,76,77),
+			          List(5,14,23,32,41,50,54,55,56,57,58,60,61,62,66,67,68,75,76,77),
+			          List(6,15,24,33,42,51,54,55,56,57,58,59,61,62,69,70,71,78,79,80),
+			          List(7,16,25,34,43,52,54,55,56,57,58,59,60,62,69,70,71,78,79,80),
+			          List(8,17,26,35,44,53,54,55,56,57,58,59,60,61,69,70,71,78,79,80),
+			          List(0,9,18,27,36,45,54,55,56,64,65,66,67,68,69,70,71,72,73,74),
+			          List(1,10,19,28,37,46,54,55,56,63,65,66,67,68,69,70,71,72,73,74),
+			          List(2,11,20,29,38,47,54,55,56,63,64,66,67,68,69,70,71,72,73,74),
+			          List(3,12,21,30,39,48,57,58,59,63,64,65,67,68,69,70,71,75,76,77),
+			          List(4,13,22,31,40,49,57,58,59,63,64,65,66,68,69,70,71,75,76,77),
+			          List(5,14,23,32,41,50,57,58,59,63,64,65,66,67,69,70,71,75,76,77),
+			          List(6,15,24,33,42,51,60,61,62,63,64,65,66,67,68,70,71,78,79,80),
+			          List(7,16,25,34,43,52,60,61,62,63,64,65,66,67,68,69,71,78,79,80),
+			          List(8,17,26,35,44,53,60,61,62,63,64,65,66,67,68,69,70,78,79,80),
+			          List(0,9,18,27,36,45,54,55,56,63,64,65,73,74,75,76,77,78,79,80),
+			          List(1,10,19,28,37,46,54,55,56,63,64,65,72,74,75,76,77,78,79,80),
+			          List(2,11,20,29,38,47,54,55,56,63,64,65,72,73,75,76,77,78,79,80),
+			          List(3,12,21,30,39,48,57,58,59,66,67,68,72,73,74,76,77,78,79,80),
+			          List(4,13,22,31,40,49,57,58,59,66,67,68,72,73,74,75,77,78,79,80),
+			          List(5,14,23,32,41,50,57,58,59,66,67,68,72,73,74,75,76,78,79,80),
+			          List(6,15,24,33,42,51,60,61,62,69,70,71,72,73,74,75,76,77,79,80),
+			          List(7,16,25,34,43,52,60,61,62,69,70,71,72,73,74,75,76,77,78,80),
+			          List(8,17,26,35,44,53,60,61,62,69,70,71,72,73,74,75,76,77,78,79) )
 
-	// Function to subset a sudoku (sdf) by a dimension (1=row, 2=col, 3=box) and value
-	def filterDim (sdf:Array[Array[Int]], dim:Int, value:Int) : Array[Array[Int]] = {
-		var out = Array.ofDim[Int](9, 4)
-		var counter = 0;
+
+
+	// Function to print sudoku array in somewhat readable fashion
+	def printSudoku (sudoku:ListBuffer[Int]) = {
+		val rowBlock = "+ - - - + - - - + - - - +"
+		var v = sudoku
+		for(r <- 0.to(12)) {
+			if(List(0, 4, 8, 12).contains(r)) print(rowBlock + "\n")
+			else {
+				for(c <- 0.to(12)) {
+					if(List(0, 4, 8).contains(c)) print("| ")
+					else if(c == 12) print("| \n")
+					else {
+						if(v(0) > 0) print(v(0) + " ")
+					    else print("  ")
+						v = v.drop(1)
+					}
+				}
+			}
+		}
+	}
+
+	// Function to get candidates for element
+	def getCandidates(sudoku:ListBuffer[Int], peers:List[List[Int]]): ListBuffer[ListBuffer[Int]] = {
+		var out = ListBuffer.fill(81)(ListBuffer.fill(1)(0))
 		for(i <- 0.to(80)) {
-			if(sdf(i)(dim) == value) {
-				out(counter) = sdf(i)
-		        counter += 1
-			} 
+			if(sudoku(i) > 0) {
+				out(i) = ListBuffer(sudoku(i))
+			} else {
+				var options = ListBuffer(1,2,3,4,5,6,7,8,9)
+				for(j <- 0.to(19)) {
+					if(sudoku(peers(i)(j)) > 0 && options.contains(sudoku(peers(i)(j)))) {
+						options -= sudoku(peers(i)(j))
+					}
+				}
+				out(i) = options
+			}
 		}
 		return out
 	}
 
-	// Function to return list of numbers already in a given dimension (1=row, 2=col, 3=box)
-	def inDim (sdf:Array[Array[Int]], dim:Int, value:Int) : ListBuffer[Int] = {
-		var inAlready = ListBuffer[Int]()
-		var i:Int = 0
-		while(i <= 80) {
-			if(sdf(i)(dim) == value & sdf(i)(0) != 0) {
-				inAlready += sdf(i)(0)
-			} 
-			i += 1
-		}
-		return inAlready;
-	}
 
-	// Function to return a List of what an element can't be
-	def cantBesGetter (sdf:Array[Array[Int]], row:Int, col:Int, box:Int) : ListBuffer[Int] = {
-		var inrow = inDim(sdf, 1, row)
-		var incol = inDim(sdf, 2, col)
-		var inbox = inDim(sdf, 3, box)
-	    return (inrow ++ incol ++ inbox).distinct
-	}
-
-	// Function to get number of unsolved elements in a sudoku
-	def numEmpties (sdf:Array[Array[Int]]) : Int = {
-		var counter = 0;
-		for(i <- 0.to(80)) {
-			if(sdf(i)(0) == 0) {
-				counter += 1
-			}
-		}
-		return counter
-	}
-
-	// Function to populate an element if there is only one number it can't be
-	def populateCantBe (sdf:Array[Array[Int]]) : Array[Array[Int]] = {
-		for(i <- 0.to(80)) {
-			if(sdf(i)(0) == 0) {
-				var cantbes = cantBesGetter(sdf, sdf(i)(1), sdf(i)(2), sdf(i)(3))
-				if(cantbes.size == 8) {
-					sdf(i)(0) = nums.diff(cantbes)(0)
-				}
-			}
-		}
-		return sdf
-	}
-
-	// Function to populate an element if everything else in its dimension can't be a given number
-	def populateExclusive (sdf:Array[Array[Int]], element:Int, dim:Int) : Array[Array[Int]] = {
-
-		// Get all numbers that can't be elsewhere in the dimension
-		var value = sdf(element)(dim)
-		var inAlready = inDim(sdf, dim, value)
-		if(inAlready.size == 8) {
-			sdf(element)(0) = nums.diff(inAlready)(0)
-		}
+	// Function to get which elements are empty
+	def getEmpties(sudoku:ListBuffer[Int]): ListBuffer[Int] = {
 		var out = ListBuffer[Int]()
 		for(i <- 0.to(80)) {
-			if(sdf(i)(dim) == value & sdf(i)(0) == 0 & i != element) {
-				var cantBes = cantBesGetter(sdf, sdf(i)(1), sdf(i)(2), sdf(i)(3))
-				cantBes = cantBes.diff(inAlready)
-				for(j <- 0.to(cantBes.size - 1)) {
-					out += cantBes(j)
-				}
-			}
+			if(sudoku(i) == 0) out += i
 		}
-
-		// If there's one number that can't be in any of the other open slots,
-		// populate this element with it.
-		if(out.size > 0) {
-			var mapped = out.groupBy(identity).mapValues(_.size)
-			for(j <- out.distinct) {
-				if(mapped(j) == (8 - inAlready.size)) {
-					sdf(element)(0) = j
-				}
-			}
-		}
-		return sdf
+		return out
 	}
 
-	// Function to solve with pure sudoku logic
-	def solveLogic (sdf:Array[Array[Int]]) : Array[Array[Int]] = {
-		var blanksStart = numEmpties(sdf)
-		if(blanksStart == 0) {
-			return sdf
+	// Function to get the number of candidates for each empty element
+	def getLengths(empties:ListBuffer[Int], candidates:ListBuffer[ListBuffer[Int]]): ListBuffer[Int] = {
+		var out = ListBuffer.fill(empties.length)(0)
+		for(i <- 0.to(empties.length-1)) {
+			out(i) = candidates(empties(i)).length
 		}
-		var blanksFinish = 1
-		while(blanksFinish > 0 & blanksStart != blanksFinish) {
-			blanksStart = numEmpties(sdf)
-			populateCantBe(sdf)
-			for(i <- 0.to(80)) {
-				if(sdf(i)(0) == 0) {
-					for(j <- 1.to(3)) {
-				    	if(sdf(i)(0) == 0) {
-			    	    	populateExclusive(sdf, i, j)
-				        }
-				    }
-				    blanksFinish = numEmpties(sdf)
-				}
-			}
-		}
-		return sdf
+		return out
 	}
 
-	// Function to solve with backtracking
-	def solveBacktracking (sdf:Array[Array[Int]], emptyElements:ListBuffer[Int]) : Boolean = {
+
+	// Function to sort empties by their number of candidates (ascending)
+	def sortEmpties(empties:ListBuffer[Int], lengths:ListBuffer[Int]): ListBuffer[Int] = {
+		val sortedEmpties = (empties zip lengths).sortBy(_._2).unzip._1
+		return sortedEmpties
+	}
+
+
+	// Function to solve a puzzle by backtracking
+	def solveBacktracking(sudoku:ListBuffer[Int], empties:ListBuffer[Int], candidates:ListBuffer[ListBuffer[Int]]): Boolean = {
 		
-		if(emptyElements.size == 0) {
-			return true
-		}
+		if(empties.length == 0) return true 
 
-		var subSdf = sdf(emptyElements(0))
-	    var cantBes = cantBesGetter(sdf, subSdf(1), subSdf(2), subSdf(3))
-	    var options = nums.toList.diff(cantBes)
-	    for(i <- options) {
-	    	sdf(emptyElements(0))(0) = i
-	    	if(solveBacktracking(sdf, emptyElements.slice(1, emptyElements.size))) {
-	    		return true
-	    	} else {
-	    		sdf(emptyElements(0))(0) = 0
-	    	} 
-	    }
-	    return false
-	}
-
-	// Function to print sudoku array in somewhat readable fashion
-	def printSudoku (sdf:Array[Array[Int]]) = {
-		for(i <- nums) {
-			var subSdf = filterDim(sdf, 1, i)
-			for(j <- 0.to(8)) {
-				print(subSdf(j)(0) + " ")
+		var index = empties(0)
+		var clen = candidates(index).length 
+		if(clen > 1) {
+			var i = 0
+			while(i < empties.length && clen > 1) {
+				if(candidates(empties(i)).length < clen) {
+					clen = candidates(empties(i)).length
+					index = empties(i)
+				}
+				i += 1
 			}
-			print("\n")
 		}
-	}
 
+		var options = candidates(index)
+		for(o <- options) {
+			var to_update = ListBuffer[Int]()
+			var j = 0
+			while(j < 20) {
+				var p = peers(index)(j)
+				if(candidates(p).length > 1) {
+					if(candidates(p).contains(o)) {
+						candidates(p) -= o
+						to_update += p
+					}
+				}
+				j += 1
+			} 
 
-	// Full solver.  First try logic, then backtracking
-	def solveSudoku (sdf:Array[Array[Int]]) : Array[Array[Int]] = {
-		solveLogic(sdf:Array[Array[Int]])
-		if(numEmpties(sdf) != 0) {
-			var emptyElements = ListBuffer[Int]()
-			for(i <- 0.to(80) if sdf(i)(0) == 0) {
-				emptyElements += i
+			sudoku(index) = o
+			empties -= index
+			if(solveBacktracking(sudoku, empties, candidates)) return true
+			sudoku(index) = 0
+			empties += index 
+
+			for(p <- to_update) {
+				candidates(p) += o
 			}
-			solveBacktracking(sdf, emptyElements)
 		}
-		printSudoku(sdf)
-		return sdf
+		return false
 	}
 
-
-
-
-	// EXAMPLE
-
-	// Board metadata
-	var rows = List(1, 2, 3, 4, 5, 6, 7, 8, 9, 
-					1, 2, 3, 4, 5, 6, 7, 8, 9, 
-					1, 2, 3, 4, 5, 6, 7, 8, 9,
-					1, 2, 3, 4, 5, 6, 7, 8, 9,
-					1, 2, 3, 4, 5, 6, 7, 8, 9,
-					1, 2, 3, 4, 5, 6, 7, 8, 9,
-					1, 2, 3, 4, 5, 6, 7, 8, 9, 
-					1, 2, 3, 4, 5, 6, 7, 8, 9, 
-					1, 2, 3, 4, 5, 6, 7, 8, 9)
-	var cols = rows.sorted
-	var boxs = List(1, 1, 1, 4, 4, 4, 7, 7, 7, 
-					1, 1, 1, 4, 4, 4, 7, 7, 7, 
-					1, 1, 1, 4, 4, 4, 7, 7, 7, 
-					2, 2, 2, 5, 5, 5, 8, 8, 8, 
-					2, 2, 2, 5, 5, 5, 8, 8, 8, 
-					2, 2, 2, 5, 5, 5, 8, 8, 8, 
-					3, 3, 3, 6, 6, 6, 9, 9, 9, 
-					3, 3, 3, 6, 6, 6, 9, 9, 9, 
-					3, 3, 3, 6, 6, 6, 9, 9, 9)
-
-
-
-	// Sudoku values
-	var vals = List(2, 4, 0, 0, 8, 0, 5, 0, 0, 
-					1, 0, 6, 0, 0, 5, 0, 2, 0, 
-					0, 8, 5, 2, 7, 0, 0, 0, 0, 
-					0, 0, 0, 5, 0, 2, 0, 7, 0, 
-					0, 0, 2, 0, 0, 0, 1, 0, 0, 
-					0, 1, 0, 3, 0, 4, 0, 0, 0, 
-					0, 0, 0, 0, 5, 7, 4, 6, 0, 
-					0, 2, 0, 9, 0, 0, 7, 0, 8, 
-					0, 0, 4, 0, 2, 0, 0, 1, 9)
-
-	var sudoku = Array.ofDim[Int](81, 4)
-	for(i <- 0.to(80)) {
-	  sudoku(i)(0) = vals(i)
-	  sudoku(i)(1) = rows(i)
-	  sudoku(i)(2) = cols(i)
-	  sudoku(i)(3) = boxs(i)
+	// Full sudoku solver
+	def solveSudoku(sudoku:ListBuffer[Int]): ListBuffer[Int] = { 
+		var empties = getEmpties(sudoku)
+		var candidates = getCandidates(sudoku, peers)
+		var lengths = getLengths(empties, candidates)
+		empties = sortEmpties(empties, lengths)
+		solveBacktracking(sudoku, empties, candidates)
+		return sudoku;
 	}
 
+	// Some sample puzzles
+	var easy_sudoku = ListBuffer(2,4,0,0,8,0,5,0,0,1,0,6,0,0,5,0,2,0,0,8,5,2,7,0,0,0,0,0,0,0,5,0,2,0,7,0,0,0,2,0,0,0,1,0,0,0,1,0,3,0,4,0,0,0,0,0,0,0,5,7,4,6,0,0,2,0,9,0,0,7,0,8,0,0,4,0,2,0,0,1,9)
+	var hard_sudoku = ListBuffer(8,0,0,0,0,0,2,0,0,7,0,0,0,8,0,0,5,9,0,0,0,9,0,2,0,0,6,5,0,0,0,7,0,0,1,0,0,0,4,0,0,8,0,0,0,0,8,0,0,0,6,0,0,0,0,0,0,0,0,0,0,2,0,0,0,0,0,1,0,0,3,0,3,7,0,0,6,0,0,0,0)
+	var evil_sudoku = ListBuffer(0,0,0,0,0,0,0,1,2,0,0,0,0,0,0,0,0,3,0,0,2,3,0,0,4,0,0,0,0,1,8,0,0,0,0,5,0,6,0,0,7,0,8,0,0,0,0,0,0,0,9,0,0,0,0,0,8,5,0,0,0,0,0,9,0,0,0,4,0,5,0,0,4,7,0,0,0,6,0,0,0)
 
 	// Time a Function
 	def time[R](block: => R): R = {
@@ -228,13 +226,31 @@ object Sudoku {
 	    return result
 	}
 
-	// Solves in about 1/100th of a second on macbook air (2X the time of cpp version)
+	// Execute
 	def main(args:Array[String]) : Unit = {
-		printSudoku(sudoku)
-		print("\n")
+		print("easy sudoku\n")
+		printSudoku(easy_sudoku)
 		time {
-			solveSudoku(sudoku)
+			easy_sudoku = solveSudoku(easy_sudoku)
 		}
+		printSudoku(easy_sudoku)
+		print("\n")
+
+		print("hard sudoku\n")
+		printSudoku(hard_sudoku)
+		time {
+			hard_sudoku = solveSudoku(hard_sudoku)
+		}
+		printSudoku(hard_sudoku)
+		print("\n")
+
+		print("evil sudoku\n")
+		printSudoku(evil_sudoku)
+		time {
+			evil_sudoku = solveSudoku(evil_sudoku)
+		}
+		printSudoku(evil_sudoku)
+		print("\n")
 	}
 
 }
